add navigation w react router & install firebase
